Extract stats player replay source helpers
